+++ java/com/turbovnc/rfb/Params.java
@@ -1,6 +1,6 @@
 /* Copyright (C) 2002-2005 RealVNC Ltd.  All Rights Reserved.
  * Copyright 2004-2005 Cendio AB.
- * Copyright (C) 2012-2018, 2020-2021 D. R. Commander.  All Rights Reserved.
+ * Copyright (C) 2012-2018, 2020-2022 D. R. Commander.  All Rights Reserved.
  * Copyright (C) 2011-2012, 2016 Brian P. Hinz
  *
  * This is free software; you can redistribute it and/or modify
@@ -382,8 +382,8 @@ public final class Params {
 
   public static IntParameter maxClipboard =
   new IntParameter("MaxClipboard",
-  "Maximum permitted length of an outgoing clipboard update (in bytes)",
-  1048576);
+  "Maximum permitted length of an incoming or outgoing clipboard update (in " +
+  "bytes)", 1048576);
 
   public static BoolParameter noNewConn =
   new BoolParameter("NoNewConn",